Added accessors for Form and Frame to ModalForm
This allows lazy styling of the textbox to occur since the style
settings in K9S are decoupled from the theme settings in tcell. Without
these accessors we're unable to override several style choices that
occur after the stock style is set but before the K9S style is set.