Add ModalStyleOpts to allow easier styling of ModalForms
In K9S, almost all forms should have the same look and feel. This was
previously acomplished by applying the theme settings right before the
form was created.

By moving the style to package variable, defaultStyle, this allows
ModalForms to be shown by default (ie with legacy behavior) and have the
correct style while still allowing individual dialogs to be shown with a
different look and feel (for example an error dialog) via DrawWithStyle.